+++ internal/js_printer/js_printer_test.go
@@ -374,13 +374,13 @@ func TestString(t *testing.T) {
 }
 
 func TestTemplate(t *testing.T) {
-	expectPrinted(t, "let x = `\\1`", "let x = `\x01`;\n")
+	expectPrinted(t, "let x = `\\0`", "let x = `\\0`;\n")
 	expectPrinted(t, "let x = `\\x01`", "let x = `\x01`;\n")
-	expectPrinted(t, "let x = `\\1${0}`", "let x = `\x01${0}`;\n")
+	expectPrinted(t, "let x = `\\0${0}`", "let x = `\\0${0}`;\n")
 	expectPrinted(t, "let x = `\\x01${0}`", "let x = `\x01${0}`;\n")
-	expectPrinted(t, "let x = `${0}\\1`", "let x = `${0}\x01`;\n")
+	expectPrinted(t, "let x = `${0}\\0`", "let x = `${0}\\0`;\n")
 	expectPrinted(t, "let x = `${0}\\x01`", "let x = `${0}\x01`;\n")
-	expectPrinted(t, "let x = `${0}\\1${1}`", "let x = `${0}\x01${1}`;\n")
+	expectPrinted(t, "let x = `${0}\\0${1}`", "let x = `${0}\\0${1}`;\n")
 	expectPrinted(t, "let x = `${0}\\x01${1}`", "let x = `${0}\x01${1}`;\n")
 
 	expectPrinted(t, "let x = String.raw`\\1`", "let x = String.raw`\\1`;\n")